+++ example/lib/main.dart
@@ -93,6 +93,7 @@ class _MyAppState extends State<MyApp> {
   List<int> selectedItemsMultiDialogWithCountAndWrap = [];
   List<int> selectedItemsMultiDialogPaged = [];
   List<Map<String, dynamic>> selectedItemsMultiMenuPagedFuture = [];
+  List<Map<String, dynamic>> selectedItemsMultiDialogPagedFuture = [];
 
   static const String appTitle = "Search Choices demo";
   final String loremIpsum =
@@ -1255,7 +1256,7 @@ class _MyAppState extends State<MyApp> {
                           width: 1,
                         ),
                       ),
-                      margin: EdgeInsets.all(1),
+                      margin: EdgeInsets.all(10),
                       child: Padding(
                         padding: const EdgeInsets.all(6),
                         child: Text(
@@ -1412,7 +1413,7 @@ class _MyAppState extends State<MyApp> {
                           width: 1,
                         ),
                       ),
-                      margin: EdgeInsets.all(1),
+                      margin: EdgeInsets.all(12),
                       child: Padding(
                         padding: const EdgeInsets.all(6),
                         child: Text(
@@ -1562,7 +1563,7 @@ class _MyAppState extends State<MyApp> {
                           width: 1,
                         ),
                       ),
-                      margin: EdgeInsets.all(1),
+                      margin: EdgeInsets.all(10),
                       child: Padding(
                         padding: const EdgeInsets.all(6),
                         child: Text(
@@ -1598,9 +1599,147 @@ class _MyAppState extends State<MyApp> {
           });
         },
       ),
+      "Multi dialog paged future": SearchChoices.multiple(
+        futureSelectedValues: selectedItemsMultiDialogPagedFuture,
+        hint: kIsWeb ? "Example not for web" : "Select capitals",
+        searchHint: "",
+        dialogBox: true,
+        onChanged: kIsWeb
+            ? null
+            : (value) {
+                setState(() {
+                  selectedItemsMultiDialogPagedFuture = value;
+                });
+              },
+        isExpanded: true,
+        itemsPerPage: 10,
+        currentPage: currentPage,
+        selectedValueWidgetFn: (item) {
+          return (Center(
+              child: Card(
+                  shape: RoundedRectangleBorder(
+                    borderRadius: BorderRadius.circular(4),
+                    side: BorderSide(
+                      color: Colors.grey,
+                      width: 1,
+                    ),
+                  ),
+                  margin: EdgeInsets.all(1),
+                  child: Padding(
+                    padding: const EdgeInsets.all(6),
+                    child: Text(item["capital"]),
+                  ))));
+        },
+        futureSearchFn: (String? keyword, String? orderBy, bool? orderAsc,
+            List<Tuple2<String, String>>? filters, int? pageNb) async {
+          String filtersString = "";
+          int i = 1;
+          filters?.forEach((element) {
+            filtersString += "&filter" +
+                i.toString() +
+                "=" +
+                element.item1 +
+                "," +
+                element.item2;
+            i++;
+          });
+          Response response = await get(Uri.parse(
+                  "https://searchchoices.jod.li/exampleList.php?page=${pageNb ?? 1},10${orderBy == null ? "" : "&order=" + orderBy + "," + (orderAsc ?? true ? "asc" : "desc")}${(keyword == null || keyword.isEmpty) ? "" : "&filter=capital,cs," + keyword}$filtersString"))
+              .timeout(Duration(
+            seconds: 10,
+          ));
+          if (response.statusCode != 200) {
+            throw Exception("failed to get data from internet");
+          }
+          dynamic data = jsonDecode(response.body);
+          int nbResults = data["results"];
+          List<DropdownMenuItem> results = (data["records"] as List<dynamic>)
+              .map<DropdownMenuItem>((item) => DropdownMenuItem(
+                    value: item,
+                    child: Card(
+                      shape: RoundedRectangleBorder(
+                        borderRadius: BorderRadius.circular(4),
+                        side: BorderSide(
+                          color: Colors.blue,
+                          width: 1,
+                        ),
+                      ),
+                      margin: EdgeInsets.all(10),
+                      child: Padding(
+                        padding: const EdgeInsets.all(6),
+                        child: Text(
+                            "${item["capital"]} - ${item["country"]} - ${item["continent"]} - pop.: ${item["population"]}"),
+                      ),
+                    ),
+                  ))
+              .toList();
+          return (Tuple2<List<DropdownMenuItem>, int>(results, nbResults));
+        },
+        futureSearchOrderOptions: {
+          "country": {
+            "icon": Wrap(children: [
+              Icon(Icons.flag),
+              Text(
+                "Country",
+              )
+            ]),
+            "asc": true
+          },
+          "capital": {
+            "icon":
+                Wrap(children: [Icon(Icons.location_city), Text("Capital")]),
+            "asc": true
+          },
+          "continent": {"icon": "Continent", "asc": true},
+          "population": {
+            "icon": Wrap(children: [Icon(Icons.people), Text("Population")]),
+            "asc": false
+          },
+        },
+        futureSearchFilterOptions: {
+          "continent": {
+            "icon": Text("Continent"),
+            "exclusive": true,
+            "values": [
+              {"eq,Africa": "Africa"},
+              {"eq,Americas": "Americas"},
+              {"eq,Asia": "Asia"},
+              {"eq,Australia": "Australia"},
+              {"eq,Europe": "Europe"},
+              {"eq,Oceania": "Oceania"}
+            ]
+          },
+          "population": {
+            "icon": Wrap(children: [Icon(Icons.people), Text("Population")]),
+            "exclusive": true,
+            "values": [
+              {
+                "lt,1000": Wrap(children: [Icon(Icons.person), Text("<1,000")])
+              },
+              {
+                "lt,100000":
+                    Wrap(children: [Icon(Icons.person_add), Text("<100,000")])
+              },
+              {
+                "lt,1000000": Wrap(
+                    children: [Icon(Icons.nature_people), Text("<1,000,000")])
+              },
+              {
+                "gt,1000000":
+                    Wrap(children: [Icon(Icons.people), Text(">1,000,000")])
+              },
+              {
+                "gt,10000000": Wrap(
+                    children: [Icon(Icons.location_city), Text(">10,000,000")])
+              },
+            ]
+          },
+        },
+      ),
     };
 
     return MaterialApp(
+      debugShowCheckedModeBanner: false,
       navigatorKey: MyApp.navKey,
       home: asTabs
           ? DefaultTabController(